Merged Participant and ParticipantVO into the same object by having VO extending from Participant and deprecating ParticipantVO.
All of those properties that are common are now delegated to Participant and the extras in VO remain there for legacy support.

ParticipantUtil is now moved to the core project since it is not dedicated to only remote tests.  Added a unit test to ensure the inheritance doesn't caase any problems (ParticipantTest).